# contracts/openapi.yaml
components:
  schemas:
    Body_import_map_api_v1_maps_import_post:
      properties:
        file:
          format: binary
          title: File
          type: string
      required:
      - file
      title: Body_import_map_api_v1_maps_import_post
      type: object
    ComponentStatus:
      enum:
      - HEALTHY
      - DEGRADED
      - UNAVAILABLE
      - DISABLED
      title: ComponentStatus
      type: string
    ExperimentCreateRequest:
      additionalProperties: false
      properties:
        map_id:
          anyOf:
          - minLength: 1
            type: string
          - type: 'null'
          title: Map Id
        scenario_id:
          format: uuid
          title: Scenario Id
          type: string
      required:
      - scenario_id
      title: ExperimentCreateRequest
      type: object
    ExperimentStatus:
      enum:
      - CREATED
      - PREPARING
      - READY
      - RUNNING
      - PAUSED
      - STOPPING
      - COMPLETED
      - FAILED
      title: ExperimentStatus
      type: string
    ExperimentView:
      additionalProperties: false
      properties:
        experiment_id:
          format: uuid
          title: Experiment Id
          type: string
        simulation_time_ms:
          minimum: 0.0
          title: Simulation Time Ms
          type: integer
        speed_multiplier:
          exclusiveMinimum: 0.0
          title: Speed Multiplier
          type: number
        status:
          $ref: '#/components/schemas/ExperimentStatus'
      required:
      - experiment_id
      - status
      - simulation_time_ms
      - speed_multiplier
      title: ExperimentView
      type: object
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    HealthResponse:
      additionalProperties: false
      properties:
        service:
          const: trafficverse-api
          default: trafficverse-api
          title: Service
          type: string
        status:
          const: ok
          default: ok
          title: Status
          type: string
      title: HealthResponse
      type: object
    MapImportJob:
      additionalProperties: false
      properties:
        error_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Code
        errors:
          default: []
          items:
            type: string
          title: Errors
          type: array
        job_id:
          format: uuid
          title: Job Id
          type: string
        map_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Map Id
        status:
          enum:
          - PENDING
          - RUNNING
          - SUCCEEDED
          - FAILED
          title: Status
          type: string
      required:
      - job_id
      - status
      title: MapImportJob
      type: object
    MapSummary:
      additionalProperties: false
      properties:
        map_id:
          minLength: 1
          title: Map Id
          type: string
        network_schema_version:
          minLength: 1
          title: Network Schema Version
          type: string
        sumo_version:
          minLength: 1
          title: Sumo Version
          type: string
        validated:
          title: Validated
          type: boolean
      required:
      - map_id
      - sumo_version
      - validated
      - network_schema_version
      title: MapSummary
      type: object
    ReadinessComponent:
      additionalProperties: false
      properties:
        component:
          minLength: 1
          title: Component
          type: string
        message:
          anyOf:
          - type: string
          - type: 'null'
          title: Message
        required:
          title: Required
          type: boolean
        status:
          $ref: '#/components/schemas/ComponentStatus'
      required:
      - component
      - status
      - required
      title: ReadinessComponent
      type: object
    ReadinessResponse:
      additionalProperties: false
      properties:
        components:
          items:
            $ref: '#/components/schemas/ReadinessComponent'
          title: Components
          type: array
        ready:
          title: Ready
          type: boolean
      required:
      - ready
      - components
      title: ReadinessResponse
      type: object
    SetSpeedRequest:
      additionalProperties: false
      properties:
        multiplier:
          exclusiveMinimum: 0.0
          maximum: 16.0
          title: Multiplier
          type: number
      required:
      - multiplier
      title: SetSpeedRequest
      type: object
    StopExperimentRequest:
      additionalProperties: false
      properties:
        reason:
          default: USER_REQUEST
          maxLength: 100
          minLength: 1
          title: Reason
          type: string
      title: StopExperimentRequest
      type: object
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
    WorkspacePageResponse:
      additionalProperties: false
      properties:
        items:
          items:
            $ref: '#/components/schemas/WorkspaceSummary'
          title: Items
          type: array
        limit:
          minimum: 1.0
          title: Limit
          type: integer
        offset:
          minimum: 0.0
          title: Offset
          type: integer
        total:
          minimum: 0.0
          title: Total
          type: integer
      required:
      - items
      - total
      - offset
      - limit
      title: WorkspacePageResponse
      type: object
    WorkspaceSummary:
      additionalProperties: false
      properties:
        created_at:
          format: date-time
          title: Created At
          type: string
        description:
          title: Description
          type: string
        name:
          maxLength: 200
          minLength: 1
          title: Name
          type: string
        updated_at:
          format: date-time
          title: Updated At
          type: string
        workspace_id:
          format: uuid
          title: Workspace Id
          type: string
      required:
      - workspace_id
      - name
      - description
      - created_at
      - updated_at
      title: WorkspaceSummary
      type: object
info:
  title: TrafficVerse API
  version: 2.1.0
openapi: 3.1.0
paths:
  /api/v1/experiments:
    post:
      operationId: create_experiment_api_v1_experiments_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExperimentCreateRequest'
        required: true
      responses:
        '202':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExperimentView'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create Experiment
  /api/v1/experiments/{experiment_id}:
    get:
      operationId: get_experiment_api_v1_experiments__experiment_id__get
      parameters:
      - in: path
        name: experiment_id
        required: true
        schema:
          format: uuid
          title: Experiment Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExperimentView'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Experiment
  /api/v1/experiments/{experiment_id}/pause:
    post:
      operationId: pause_experiment_api_v1_experiments__experiment_id__pause_post
      parameters:
      - in: path
        name: experiment_id
        required: true
        schema:
          format: uuid
          title: Experiment Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExperimentView'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Pause Experiment
  /api/v1/experiments/{experiment_id}/prepare:
    post:
      operationId: prepare_experiment_api_v1_experiments__experiment_id__prepare_post
      parameters:
      - in: path
        name: experiment_id
        required: true
        schema:
          format: uuid
          title: Experiment Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExperimentView'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Prepare Experiment
  /api/v1/experiments/{experiment_id}/resume:
    post:
      operationId: resume_experiment_api_v1_experiments__experiment_id__resume_post
      parameters:
      - in: path
        name: experiment_id
        required: true
        schema:
          format: uuid
          title: Experiment Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExperimentView'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Resume Experiment
  /api/v1/experiments/{experiment_id}/speed:
    post:
      operationId: set_speed_api_v1_experiments__experiment_id__speed_post
      parameters:
      - in: path
        name: experiment_id
        required: true
        schema:
          format: uuid
          title: Experiment Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetSpeedRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExperimentView'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Set Speed
  /api/v1/experiments/{experiment_id}/start:
    post:
      operationId: start_experiment_api_v1_experiments__experiment_id__start_post
      parameters:
      - in: path
        name: experiment_id
        required: true
        schema:
          format: uuid
          title: Experiment Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExperimentView'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Start Experiment
  /api/v1/experiments/{experiment_id}/stop:
    post:
      operationId: stop_experiment_api_v1_experiments__experiment_id__stop_post
      parameters:
      - in: path
        name: experiment_id
        required: true
        schema:
          format: uuid
          title: Experiment Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StopExperimentRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExperimentView'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Stop Experiment
  /api/v1/health:
    get:
      operationId: health_api_v1_health_get
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
          description: Successful Response
      summary: Health
  /api/v1/maps:
    get:
      operationId: maps_api_v1_maps_get
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/MapSummary'
                title: Response Maps Api V1 Maps Get
                type: array
          description: Successful Response
      summary: Maps
  /api/v1/maps/import:
    post:
      operationId: import_map_api_v1_maps_import_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_import_map_api_v1_maps_import_post'
        required: true
      responses:
        '202':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MapImportJob'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Import Map
  /api/v1/maps/import/{job_id}:
    get:
      operationId: import_status_api_v1_maps_import__job_id__get
      parameters:
      - in: path
        name: job_id
        required: true
        schema:
          format: uuid
          title: Job Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MapImportJob'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Import Status
  /api/v1/maps/{map_id}/manifest:
    get:
      operationId: map_manifest_api_v1_maps__map_id__manifest_get
      parameters:
      - in: path
        name: map_id
        required: true
        schema:
          title: Map Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                title: Response Map Manifest Api V1 Maps  Map Id  Manifest Get
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Map Manifest
  /api/v1/maps/{map_id}/network:
    get:
      operationId: map_network_api_v1_maps__map_id__network_get
      parameters:
      - in: path
        name: map_id
        required: true
        schema:
          title: Map Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Map Network
  /api/v1/ready:
    get:
      operationId: ready_api_v1_ready_get
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadinessResponse'
          description: Successful Response
      summary: Ready
  /api/v1/workspaces:
    get:
      operationId: list_workspaces_api_v1_workspaces_get
      parameters:
      - in: query
        name: q
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Q
      - in: query
        name: offset
        required: false
        schema:
          default: 0
          minimum: 0
          title: Offset
          type: integer
      - in: query
        name: limit
        required: false
        schema:
          default: 50
          maximum: 200
          minimum: 1
          title: Limit
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspacePageResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List Workspaces
  /api/v1/workspaces/{workspace_id}:
    get:
      operationId: get_workspace_api_v1_workspaces__workspace_id__get
      parameters:
      - in: path
        name: workspace_id
        required: true
        schema:
          format: uuid
          title: Workspace Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceSummary'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Workspace